backend/pyproject.toml
[project]
name = "deepagent"
version = "0.1.0"
description = "DeepAgent: LangChain DeepAgents + LangGraph with MCP/Skills integration"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.6.0",
  "langchain>=0.3.7",
  "langgraph>=0.2.30",
  "langgraph-checkpoint-sqlite>=1.0.0",
  "deepagents>=0.1.0",
  "zhipuai>=2.0.0",
  "langchain-openai>=0.1.22",
  "aiofiles>=23.2.1",
  "httpx>=0.25.0",
  "starlette>=0.37.0",
  "typing-extensions>=4.9.0",
  "mcp>=1.0.0",
  "PyYAML>=6.0.1",
]

[tool.uv]
dev-dependencies = [
  "ruff>=0.3.0",
  "mypy>=1.8.0",
  "pytest>=8.0.0",
  "types-PyYAML>=6.0.12.20241230",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = false
warn_unused_configs = true
ignore_missing_imports = true
